fix: enable region to update when selecting markers or selection location  for storing marker
